+++ kotlin_sandi/adventOfCode2020/.gitignore
@@ -1,0 +1,9 @@
+# IntelliJ
+*.iml
+.idea/workspace.xml
+.idea/tasks.xml
+.idea/gradle.xml
+.idea/assetWizardSettings.xml
+.idea/dictionaries
+.idea/libraries
+.idea/caches
