+++ kotlin_sandi/adventOfCode2020/src/advents/Day2.kt
@@ -1,0 +1,2 @@
+package advents.day2
+
