+++ kotlin_sandi/adventOfCode2020/src/utility/FileUtility.kt
@@ -1,0 +1,2 @@
+package utility
+
